[css-grid] Skip simplified layout if the Grid is dirty.
Bug report bisected to an improvement in the LayoutNG OOF cache.
https://chromium-review.googlesource.com/c/chromium/src/+/1772524

There was a larger underlying issue that a LayoutGrid could have
simplified layout performed, while having a non-existent grid.

This manifested in an incorrect paint output.

This patch simply disables simplified-layout whenever there is dirty
grid. It doesn't seem the invalidation for this grid data-structure
is optimal.

Bug: 1040069
Change-Id: I347630f3fc5c330ae57b06219d1e32b6269177d2